+++ src/Oro/Bundle/FrontendBundle/Resources/public/default/js/app/service/modals-scroll-locker.js
@@ -19,6 +19,8 @@ export default () => {
         },
         // modals
         'modal:open': modal => scrollLocker.addLocker(modal.cid),
-        'modal:close': modal => scrollLocker.removeLocker(modal.cid)
+        'modal:close': modal => scrollLocker.removeLocker(modal.cid),
+        'fullscreen:popup:show': modal => scrollLocker.addLocker(modal.cid),
+        'fullscreen:popup:close': modal => scrollLocker.removeLocker(modal.cid)
     });
 };