+++ src/Oro/Bundle/FrontendBundle/Resources/public/default/js/app/views/frontend-abstract-filter.js
@@ -2,12 +2,16 @@ define(function(require, exports, module) {
     'use strict';
 
     const _ = require('underscore');
+    const __ = require('orotranslation/js/translator');
     const AbstractFilter = require('oro/filter/abstract-filter');
 
     let config = require('module-config').default(module.id);
 
     config = _.extend({
-        animationDuration: 0
+        animationDuration: 0,
+        clearFilterSelector: '[data-role="clear-filter"]',
+        filterEnableValueBadge: true,
+        allowClearButtonInFilter: true
     }, config);
 
     const FrontendAbstractFilter = AbstractFilter.extend({
@@ -17,6 +21,28 @@ define(function(require, exports, module) {
          * @property {Number}
          */
         animationDuration: config.animationDuration,
+
+        /**
+        * Reset filter selector
+        *
+        * @property {string}
+        */
+        clearFilterSelector: config.clearFilterSelector,
+
+        /**
+         * Enable showing badge with count of selected values
+         *
+         * @property {boolean}
+         */
+        filterEnableValueBadge: config.filterEnableValueBadge,
+
+        /**
+         * Enable reset button for particular filter
+         * Allow reset filter separately
+         *
+         * @property {boolean}
+         */
+        allowClearButtonInFilter: config.allowClearButtonInFilter,
 
         /**
          * @inheritdoc
@@ -48,6 +74,44 @@ define(function(require, exports, module) {
                     element.parent().removeClass(this.buttonActiveClass);
                 });
             }
+        },
+
+        /**
+         * @return {Object}
+         */
+        getTemplateDataProps() {
+            return {
+                ...FrontendAbstractFilter.__super__.getTemplateDataProps.call(this),
+                allowClearButtonInFilter: this.allowClearButtonInFilter,
+                clearFilterButtonAriaLabel: __('oro.filter.clearFilterButton.aria_label', {
+                    label: `${__('oro.filter.by')} ${this.label}`}
+                )
+            };
+        },
+
+        getHintChips() {
+            return this.subview('hint').getChips();
+        },
+
+        _setInputValue(input, value) {
+            const $input = this.$(input);
+
+            switch ($input.attr('type')) {
+                case 'radio':
+                    $input.each((index, input) => {
+                        const $input = this.$(input);
+                        if ($input.attr('value') === value) {
+                            $input.prop('checked', true).trigger('change');
+                        } else {
+                            $input.prop('checked', false);
+                        }
+                    });
+                    break;
+                default:
+                    $input.val(value);
+            }
+
+            return this;
         }
     });