+++ src/Oro/Bundle/FrontendBundle/Resources/public/default/js/app/views/frontend-select-filter.js
@@ -2,13 +2,17 @@ define(function(require, exports, module) {
     'use strict';
 
     const _ = require('underscore');
+    const __ = require('orotranslation/js/translator');
     const SelectFilter = require('oro/filter/select-filter');
     const MultiselectDecorator = require('orofrontend/js/app/datafilter/frontend-multiselect-decorator');
+    const FilterBadgeHintView = require('orofrontend/default/js/app/views/filter-badge-hint-view').default;
     const FilterCountHelper = require('orofrontend/js/app/filter-count-helper');
+    const tools = require('oroui/js/tools');
     let config = require('module-config').default(module.id);
 
     config = _.extend({
-        closeAfterChose: true
+        closeAfterChose: true,
+        minimumResultsForSearch: 7
     }, config);
 
     const FrontendSelectFilter = SelectFilter.extend(_.extend({}, FilterCountHelper, {
@@ -16,6 +20,11 @@ define(function(require, exports, module) {
          * @property
          */
         closeAfterChose: config.closeAfterChose,
+
+        /**
+         * @property
+         */
+        minimumResultsForSearch: config.minimumResultsForSearch,
 
         /**
          * @property
@@ -52,14 +61,31 @@ define(function(require, exports, module) {
         listen: {
             'metadata-loaded': 'onMetadataLoaded',
             'total-records-count-updated': 'onTotalRecordsCountUpdate',
-            'filters-manager:after-applying-state mediator': 'rerenderFilter'
+            'filters-manager:after-applying-state mediator': 'rerenderFilter',
+            'change': 'onChangeFilter'
         },
 
         /**
          * @inheritdoc
          */
         constructor: function FrontendSelectFilter(options) {
+            this.onChangeFilter = _.debounce(this.onChangeFilter.bind(this));
             FrontendSelectFilter.__super__.constructor.call(this, options);
+        },
+
+        rendered() {
+            if (this.subview('filter:badge-hint')) {
+                this.subview('filter:badge-hint').dispose();
+            }
+
+            if (this.filterEnableValueBadge) {
+                this.subview('filter:badge-hint', new FilterBadgeHintView({
+                    filter: this,
+                    container: this.$('.filter-criteria-selector')
+                }));
+            }
+
+            return FrontendSelectFilter.__super__.rendered.call(this);
         },
 
         /**
@@ -84,8 +110,19 @@ define(function(require, exports, module) {
          */
         _initializeSelectWidget() {
             this.widgetOptions = Object.assign({}, this.widgetOptions, {
-                additionalClass: !this.isToggleMode()
+                additionalClass: !this.isToggleMode(),
+                resetButton: this.allowClearButtonInFilter ? {
+                    label: __('oro.filter.clearFilterButton.text'),
+                    attr: {
+                        'class': 'btn btn--flat filter-clear hidden',
+                        'aria-label': __('oro.filter.clearFilterButton.aria_label', {
+                            label: `${__('oro.filter.by')} ${this.label}`}
+                        )
+                    },
+                    onClick: this.onClickSelectWidgetResetButton.bind(this)
+                } : null
             });
+            this.contextSearch = Array.isArray(this.choices) && this.choices.length > this.minimumResultsForSearch;
 
             return FrontendSelectFilter.__super__._initializeSelectWidget.call(this);
         },
@@ -104,6 +141,8 @@ define(function(require, exports, module) {
             } else {
                 FrontendSelectFilter.__super__._onClickFilterArea.call(this, e);
             }
+
+            this.toggleVisibilityClearFilterButton();
         },
 
         toggleFilter: function() {
@@ -146,6 +185,23 @@ define(function(require, exports, module) {
 
         isToggleMode: function() {
             return this.renderMode === 'toggle-mode';
+        },
+
+        onClickSelectWidgetResetButton() {
+            this.reset();
+            this.toggleVisibilityClearFilterButton();
+        },
+
+        onChangeFilter() {
+            this.toggleVisibilityClearFilterButton();
+        },
+
+        toggleVisibilityClearFilterButton(hidden) {
+            if (hidden === void 0) {
+                hidden = tools.isEqualsLoosely(this.getValue(), this.emptyValue);
+            }
+
+            this.selectWidget && this.selectWidget.toggleVisibilityResetButton(hidden);
         }
     }));