+++ src/Oro/Bundle/FrontendBundle/Resources/public/js/app/datafilter/frontend-collection-filters-manager.js
@@ -6,6 +6,9 @@ define(function(require, exports, module) {
     const __ = require('orotranslation/js/translator');
     const CollectionFiltersManager = require('orofilter/js/collection-filters-manager');
     const MultiselectDecorator = require('orofrontend/js/app/datafilter/frontend-manage-filters-decorator');
+    const FrontendCollapsableHintsView = require('./frontend-collapsable-hints-view').default;
+    const ScrollShadowView = require('orofrontend/js/app/views/scroll-shadow-view').default;
+
     let config = require('module-config').default(module.id);
     config = _.extend({
         templateData: {
@@ -26,6 +29,10 @@ define(function(require, exports, module) {
          * @inheritdoc
          */
         enableMultiselectWidget: true,
+
+        multiselectResetButtonLabel: __('oro_frontend.filter_manager.resetFilter'),
+
+        enableScrollContainerShadow: false,
 
         /**
          * @inheritdoc
@@ -51,7 +58,12 @@ define(function(require, exports, module) {
          */
         templateData: config.templateData,
 
-        optionNames: CollectionFiltersManager.prototype.optionNames.concat(['fullscreenTemplate']),
+        optionNames: CollectionFiltersManager.prototype.optionNames.concat([
+            'fullscreenTemplate', 'filtersStateElement', 'filterEnableValueBadge', 'allowClearButtonInFilter',
+            'hintsToggledStatus', 'enableScrollContainerShadow', 'closeFilterManagerOnOutClick'
+        ]),
+
+        hintsExpanded: false,
 
         /**
          * @inheritdoc
@@ -60,11 +72,24 @@ define(function(require, exports, module) {
             FrontendCollectionFiltersManager.__super__.constructor.call(this, options);
         },
 
+        initialize(options) {
+            if (this.closeFilterManagerOnOutClick !== void 0) {
+                this.multiselectParameters.closeOnOutOfClick = this.closeFilterManagerOnOutClick;
+            }
+
+            FrontendCollectionFiltersManager.__super__.initialize.call(this, options);
+        },
+
         /**
          * @inheritdoc
          */
         render: function() {
             FrontendCollectionFiltersManager.__super__.render.call(this);
+
+            if (this.filtersStateElement && this.filtersStateElement instanceof $) {
+                this.filtersStateElement.remove();
+            }
+
             this.finallyOfRender();
             return this;
         },
@@ -83,7 +108,8 @@ define(function(require, exports, module) {
                 })
                 .prepend(_.macros('oroui::renderIcon')({
                     name: 'settings'
-                }));
+                }))
+                .tooltip();
         },
 
         /**
@@ -94,11 +120,15 @@ define(function(require, exports, module) {
          */
         _createButtonReset: function() {
             // Use link to keep focus even on disabled state
+            const icon = _.macros('oroui::renderIcon')({
+                name: 'undo'
+            });
+
             return $(`
                 <div class="datagrid-manager__footer">
-                    <a href="#" role="button" class="btn btn--no-padding"
+                    <a href="#" role="button" class="btn"
                         data-role="reset-filters">
-                        <span class="fa-refresh" aria-hidden="true"></span>${this.multiselectResetButtonLabel}
+                        ${icon}${this.multiselectResetButtonLabel}
                     </a>
                 </div>
             `);
@@ -128,9 +158,45 @@ define(function(require, exports, module) {
             }
         },
 
+        _onFilterUpdated() {
+            FrontendCollectionFiltersManager.__super__._onFilterUpdated.call(this);
+            this.subview('collapsableHints') && this.subview('collapsableHints').update();
+        },
+
+        _onFilterChanged() {
+            FrontendCollectionFiltersManager.__super__._onFilterChanged.call(this);
+            this.subview('collapsableHints') && this.subview('collapsableHints').update();
+        },
+
+        _onFilterDisabled(filter) {
+            FrontendCollectionFiltersManager.__super__._onFilterDisabled.call(this, filter);
+            this.subview('collapsableHints') && this.subview('collapsableHints').update();
+        },
+
         finallyOfRender: function() {
             if (this.$el.data('layout') === 'separate') {
                 this.initLayout();
+            }
+
+            this.subview('collapsableHints', new FrontendCollapsableHintsView({
+                autoRender: true,
+                filterManager: this,
+                filters: this.filters,
+                container: this.getHintContainer(),
+                toggled: this.hintsExpanded
+            }));
+
+            this.listenTo(
+                this.subview('collapsableHints'),
+                'hints:change-visibility',
+                toggledStatus => this.hintsExpanded = toggledStatus
+            );
+
+            if (this.renderMode === 'toggle-mode' && this.enableScrollContainerShadow) {
+                this.subview('scroll-shadow', new ScrollShadowView({
+                    el: this.el,
+                    scrollTarget: '[data-filters-items]'
+                }));
             }
         }
     });